requirements.txt
aiohttp
aiosignal
annotated-types
anthropic
anyio
attrs
bert-score
certifi
cffi
charset-normalizer
click
cryptography
dataclasses-json
defusedxml
distro
filelock
frozenlist
fsspec
h11
httpcore
httpx
huggingface-hub
idna
jiter
joblib
jsonpatch
jsonpointer
langchain
langchain-anthropic
langchain-community
langchain-core
langchain-experimental
langchain-openai
langchain-text-splitters
langsmith
llama-index-llms-openai
llm-blender
marshmallow
multidict
mypy-extensions
nltk
numpy
openai
orjson
packaging
pandas
pdfminer.six
pdfplumber
pillow
pycparser
pydantic
pydantic_core
pypdfium2
pytesseract
python-dateutil
pytz
PyYAML
ragas==0.1.9
regex
requests
sentence-transformers==3.0.1
six
sniffio
SQLAlchemy
tenacity
tiktoken
tokenizers
torchmetrics==1.4.0.post0
tqdm
typing-inspect
typing_extensions
tzdata
urllib3
yarl
